Aparecen bolas
- Aparecen 12 bolas de color random en posiciones randomicas
- se modifica panelPrincipal, dentro de el no crea otro panel que se añade a la ventana, sino que el mismo panelPrincipal se agrega a la ventana
- esto ultimo es para facilitar la utilización de paint()

+++ Proyecto Semestral/src/proyecto/semestral/BolaBlanca.java
@@ -4,10 +4,10 @@ import java.awt.Color;
 
 /**
  * BolaBlanca
- */ 
+ */
 public class BolaBlanca extends Bola {
 
-    public BolaBlanca(double x, double y, double radio) {
+    public BolaBlanca(int x, int y, int radio) {
         super(x, y, radio, Color.white);
     }
 }
